add tests defines
