+++ tests/unittest/disaggregated/test_cluster_storage.py
@@ -96,19 +96,22 @@ class TestClusterStorage:
 
     @timeout(5)
     @pytest_async_module
-    async def test_get_prefix(self, storage_server, storage_client):
+    async def test_get_keys(self, storage_server, storage_client):
         keys = [gen_key("test_key_unique") for _ in range(3)]
-        for key in keys:
+        values = [f"test_value{i}" for i in range(3)]
+        for key, value in zip(keys, values):
             assert await storage_client.set(key,
-                                            "test_value1",
+                                            value,
                                             overwrite_if_exists=True)
 
-        answer_keys = await storage_client.get_prefix("test_key_unique")
-        assert set(keys) == set(answer_keys)
-        answer_keys = await storage_client.get_prefix(keys[0])
-        assert answer_keys == [keys[0]]
-        answer_keys = await storage_client.get_prefix(keys[1])
-        assert answer_keys == [keys[1]]
+        answer_keys = await storage_client.get_prefix("test_key_unique",
+                                                      keys_only=False)
+        assert set(keys) == set(answer_keys.keys())
+        assert set(values) == set(answer_keys.values())
+        answer_keys = await storage_client.get_prefix(keys[0], keys_only=True)
+        assert answer_keys == {keys[0]: ""}
+        answer_keys = await storage_client.get_prefix(keys[1], keys_only=True)
+        assert answer_keys == {keys[1]: ""}
 
     @pytest_ignore_tleak
     @pytest_async_module
@@ -199,7 +202,8 @@ class TestHttpClusterStorage(TestClusterStorage):
 
 
 class TestEtcdClusterStorage(TestClusterStorage):
-    __test__ = True
+    # Disable this test until Etcd functionality is ready.
+    __test__ = False
 
     @pytest.fixture(scope="class")
     def storage_server(self):
